+++ src/lib/turndownRules.ts
@@ -1,0 +1,12 @@
+import { addConfluenceRules } from './rules';
+import { Options } from 'turndown';
+
+export function getTurndownOptions(): Options {
+    return {
+        headingStyle: 'atx',
+        codeBlockStyle: 'fenced',
+        bulletListMarker: '-',
+    };
+}
+
+export { addConfluenceRules };